#7 update transaction state after complete payment
